requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-OAuthlib>=0.9.1
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Migrate>=1.3.0
FormEncode==1.3.0
SQLAlchemy==0.9.9
Unidecode==0.04.17
apikit>=0.3
celery==3.1.17
certifi>=14
cssmin==0.2.0
normality>=0.2.2
nose==1.3.4
oauth2==1.5.211
psycopg2==2.6
python-Levenshtein==0.12.0
pytz==2014.10
requests==2.6.0
six==1.9.0
tablib==0.10.0
webassets==0.10.1
wsgiref==0.1.2
Flask-Login>=0.2.11
PyYAML>=3.10
python-dateutil>=2.2
